Adds title and changes dialog name to NatDialogController
